requirements.txt
dlib==19.24.2
face-recognition==1.3.0
tqdm==4.65.0
numpy==1.24.3
scikit-learn==1.3.0
Pillow==9.5.0
opencv-python==4.8.0.74